Deck and Hand scripts
Small script to generate a full deck of cards, as well as a dynamic
hand. Creates a 52 card deck that can be used across a multitute of
games. The hand class allows a player to create a hand of n size, from
a deck of thier choice. Multiple players can use same deck, or create a
new one.